Use CSS classes to hide unavailable items
This replaces some of the toggling with a CSS class on the tools container, that when set, along with the `not-available` class on each item, set's their display property to `none`. When the button is pressed, this just toggles the container class as well.

This also updates the filtering logic to work with __SHOW AVAILABLE__. jQuery's `show` method adds inline style `display: block;`, overriding any previous logic. This again replaces jQuery's `show`/`hide` with a class, `filtered`, that is hidden with CSS.

A lot of jQuery code re-running filters, etc. get to be deleted! I'm not 100% sure splitting the display logic between the CSS/JS is a great idea though.

Closes #8